I may have made some suboptimal assumptions about how object relations should be modeled, but unfortunately it is too late to rewrite everything now. Nevertheless, I will find a way.

+++ Assignment 2/RecruitCatTreutewm/Models/JobTitle.cs
@@ -30,7 +30,7 @@ namespace RecruitCatTreutewm.Models
 
         [DisplayName("Min YOE")]
         [Range(0,100)]
-        public Nullable<decimal> RequiredYOE { get; set; 
+        public Nullable<double> RequiredYOE { get; set; 
         }
     }
 }
